requirements.txt
flask==3.0.2
cx-Oracle==8.3.0
python-dotenv==1.0.0